fixes checkbox not being cast to bool
