+++ src/modules/search/components/filtrosActivos.tsx
