+++ src/modules/search/types/searchParams.ts
@@ -7,8 +7,8 @@ export const searchSchema = z.object({
   schools: z.string().optional(),
   associations: z.string().optional(),
   type: z.enum(['fisica', 'moral']).optional(),
-  remote: z.boolean().optional(),
-  socialService: z.boolean().optional(),
+  remote: z.boolean({ coerce: true }).optional(),
+  socialService: z.boolean({ coerce: true }).optional(),
   location: z.string().optional(),
   category: z.string().optional(),
   activity: z.string().optional(),
